third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/injected-inline-style-blocked-expected.txt

CONSOLE ERROR: line 1: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'none'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 5: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'none'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

PASS 1/2
PASS 2/2
0
